Extract intensity color-mapping and PLY comment-metadata parsers
engine/src/utils/intensity.ts: getIntensityArray, hasIntensityData,
mapIntensityValue, buildIntensityColorArrayForMode - all were already
pure (no `this` dependency), just data transformation over
SpatialData/typed arrays.

engine/src/depth/commentSettings.ts: the family of getRgb24*/getPng*/
getDepth* parsers that read camera/depth params embedded as PLY
comments ("fx: 1000px", "Baseline: 50mm", etc). Three are fully pure;
the rest take a narrow DepthSettingsHost (defaultDepthSettings,
fileDepthData) for their fallback defaults.

main.ts keeps thin delegating wrappers for all of these since several
have 5-11 call sites elsewhere in the file.

main.ts: 13,286 -> 13,053 lines.

Verified: engine tsc clean, full webpack compile clean, all 59
extension tests passing under xvfb-run.
